refactor: split components in their files
